% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{boards}
\alias{boards}
\title{2019 Jeopardy! game boards}
\format{
A tibble with 13,261 rows and 8 variables:
\describe{
\item{game}{The J! Archive game ID.}
\item{round}{The round a clue is chosen.}
\item{col}{The column position left-to-right.}
\item{row}{The row position top-to-bottom.}
\item{n}{The order of clue chosen.}
\item{category}{Category title, often humorous or with instructions.}
\item{clue}{The clue read to the contestants.}
\item{answer}{The \emph{correct} answer to a clue.}
}
}
\source{
\url{https://www.j-archive.com/}
}
\usage{
boards
}
\description{
The categories, clues, and answers in the 2019 season of Jeopardy.
}
\keyword{datasets}
