% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entity-clean.R
\name{entity_clean}
\alias{entity_clean}
\title{Repair the text from J! Archive HTML}
\usage{
entity_clean(string)
}
\arguments{
\item{string}{Some text, usually from \code{\link[rvest:html_text]{rvest::html_text()}}.}
}
\value{
Normalize character vectors.
}
\description{
Repair the text from J! Archive HTML
}
