% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{players}
\alias{players}
\title{2019 Jeopardy! contestants}
\format{
A tibble with 672 rows and 6 variables:
\describe{
\item{game}{The J! Archive game ID.}
\item{show}{The sequential episode show number.}
\item{date}{The date each episodes initially aired.}
}
}
\source{
\url{https://www.j-archive.com/}
}
\usage{
players
}
\description{
The contestants in the 2019 season of Jeopardy.
}
\keyword{datasets}
