% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whatr-id.R
\name{read_game}
\alias{read_game}
\title{Scrape a J! Archive "showgame" page}
\usage{
read_game(game = NULL, date = NULL, show = NULL)
}
\arguments{
\item{game}{The J! Archive game ID number.}

\item{date}{The original date an episode aired.}

\item{show}{The sequential show number.}
}
\value{
The J! Archive showgame \code{html_document}.
}
\description{
GET the "showgame" HTML document object which can be parsed. Use with
functions like \code{\link[=whatr_clues]{whatr_clues()}}.
}
\examples{
read_game(game = 6304)
}
