% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whatr-id.R
\name{read_scores}
\alias{read_scores}
\title{Scrape a J! Archive "showscore" page}
\usage{
read_scores(game = NULL, date = NULL, show = NULL)
}
\arguments{
\item{game}{The J! Archive game ID number.}

\item{date}{The original date an episode aired.}

\item{show}{The sequential show number.}
}
\value{
The J! Archive showscore \code{html_document}.
}
\description{
GET the "showscore" HTML document object which can be parsed. Use with
functions like \code{\link[=whatr_score]{whatr_score()}}.
}
\examples{
read_scores(game = 6304)
}
