% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whatr-id.R
\name{showgame}
\alias{showgame}
\title{Scrape a J! Archive "showgame" page}
\usage{
showgame(game = NULL, date = NULL, show = NULL)
}
\arguments{
\item{game}{The J-Archive! game ID number.}

\item{date}{The original date an episode aired.}

\item{show}{The sequential show number.}
}
\value{
The J-Archive! showgame page.
}
\description{
Use \code{\link[xml2:read_html]{xml2::read_html()}} to scrape the entire page as R object.
}
