% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whatr-airdate.R
\name{whatr_airdate}
\alias{whatr_airdate}
\title{What is the air date?}
\format{A tibble with 1 row and 3 variables:
\describe{
\item{game}{The non-sequential J! Archive game ID.}
\item{show}{The sequential show number of an episode.}
\item{date}{The air date of an episode.}
}}
\usage{
whatr_airdate(game)
}
\arguments{
\item{game}{One of four types of input, all passed to \code{\link[=whatr_html]{whatr_html()}}:
\enumerate{
\item A numeric game ID.
\item Either a 'showgame' or 'showscores' HTML document.
\item A show number character starting with "#".
\item An air date like "yyyy-mm-dd".
}}
}
\value{
A tidy tibble of clue text.
}
\description{
\emph{This} date identifies when an episode was first viewed on television.
}
\examples{
whatr_airdate(game = 6304)
}
