% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whatr-answers.R
\name{whatr_answers}
\alias{whatr_answers}
\title{What are the answers?}
\format{A tibble with (usually) 61 rows and 5 variables:
\describe{
\item{round}{The round a clue is chosen.}
\item{col}{The column position left-to-right.}
\item{row}{The row position top-to-bottom.}
\item{n}{The order of clue chosen.}
\item{answer}{The \emph{correct} answer to a clue.}
}}
\usage{
whatr_answers(game)
}
\arguments{
\item{game}{Either a numeric J! Archive game ID or the HTML document object
returned by \code{\link[=whatr_html]{whatr_html()}}, which can take a game ID, air date, show
number, or another HTML document as an input. If an HTML document is not
provided, \code{\link[=whatr_html]{whatr_html()}} will be called.}
}
\value{
A tidy tibble of clue text.
}
\description{
\emph{These} must be given by the contestants in the form of a question in
response to the clues asked.
}
\examples{
whatr_answers(game = 6304)
whatr_html(6304) \%>\% whatr_answers()
}
