% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whatr-board.R
\name{whatr_board}
\alias{whatr_board}
\title{Scrape Jeopardy game board}
\usage{
whatr_board(game = NULL, date = NULL, show = NULL)
}
\arguments{
\item{game}{The J-Archive! game ID number.}

\item{date}{The original date an episode aired.}

\item{show}{The sequential show number.}
}
\value{
A tibble of categories and positions.
}
\description{
This joins together the results of \code{\link[=whatr_categories]{whatr_categories()}}, \code{\link[=whatr_clues]{whatr_clues()}} and
\link{whatr_answers} to return a single tibble with all clue/answer combinations.
}
\examples{
whatr_board(game = 6304)
}
