% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whatr-categories.R
\name{whatr_categories}
\alias{whatr_categories}
\title{What are categories?}
\format{A tibble with (usually 13) rows and 3 variables:
\describe{
\item{round}{The round a clue is chosen.}
\item{col}{The column position left-to-right.}
\item{category}{Category title, often humorous or with instructions.}
}}
\usage{
whatr_categories(game)
}
\arguments{
\item{game}{Either a numeric J! Archive game ID or the HTML document object
returned by \code{\link[=whatr_html]{whatr_html()}}, which can take a game ID, air date, show
number, or another HTML document as an input. If an HTML document is not
provided, \code{\link[=whatr_html]{whatr_html()}} will be called.}
}
\value{
Tidy tibble of category titles.
}
\description{
\emph{These} are collections of five clues related by subject matter.
}
\examples{
whatr_categories(game = 6304)
whatr_html(6304) \%>\% whatr_categories()
}
