% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whatr-clues.R
\name{whatr_clues}
\alias{whatr_clues}
\title{What are the clues?}
\format{A tibble with (usually) 61 rows and 4 variables:
\describe{
\item{round}{The round a clue is chosen.}
\item{col}{The column position left-to-right.}
\item{row}{The row position top-to-bottom.}
\item{n}{The order of clue chosen.}
\item{clue}{The clue read to the contestants.}
}}
\usage{
whatr_clues(game)
}
\arguments{
\item{game}{Either a numeric J! Archive game ID or the HTML document object
returned by \code{\link[=whatr_html]{whatr_html()}}, which can take a game ID, air date, show
number, or another HTML document as an input. If an HTML document is not
provided, \code{\link[=whatr_html]{whatr_html()}} will be called.}
}
\value{
A tidy tibble of clue text.
}
\description{
Confusingly, \emph{these} are posed to contestants in the form of an answer.
}
\examples{
whatr_clues(game = 6304)
whatr_html(6304) \%>\% whatr_clues()
}
