% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whatr-data.R
\name{whatr_data}
\alias{whatr_data}
\title{Return all a game's data from the J! Archive}
\usage{
whatr_data(game = NULL, date = NULL, show = NULL)
}
\arguments{
\item{game}{The J-Archive! game ID number.}

\item{date}{The original date an episode aired.}

\item{show}{The sequential show number.}
}
\value{
A list of tibbles each containing a data type.
}
\description{
Return all a game's data from the J! Archive
}
\examples{
whatr_data(game = 6304)
}
