% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whatr-doubles.R
\name{whatr_doubles}
\alias{whatr_doubles}
\title{Get the clue numbers that coorespond to the daily doubles}
\format{a named list
\describe{
\item{single}{The index of the daily double in the first round}
\item{double}{The indices of the two daily doubles in the second round}
}}
\usage{
whatr_doubles(html = NULL, game = NULL)
}
\arguments{
\item{html}{An HTML document from \code{\link[=read_scores]{read_scores()}}.}

\item{game}{The J-Archive! game ID number, possibly from \code{\link[=whatr_id]{whatr_id()}}.}
}
\value{
a list containing the question indices of the daily doubles
in the first and second rounds
}
\description{
Get the clue numbers that coorespond to the daily doubles
}
\examples{
whatr_doubles(game = 6304)
read_scores(6304) \%>\% whatr_douibles()
}
