% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whatr-id.R
\name{whatr_id}
\alias{whatr_id}
\title{Return a J! Archive game ID}
\usage{
whatr_id(game = NULL, date = NULL, show = NULL)
}
\arguments{
\item{game}{The J-Archive! game ID number.}

\item{date}{The original date an episode aired.}

\item{show}{The sequential show number.}
}
\value{
The J-Archive! game ID number.
}
\description{
Can return the game ID for a show air date or number.
}
