% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whatr-info.R
\name{whatr_info}
\alias{whatr_info}
\title{What is the information?}
\format{A tibble with 1 row and 3 variables:
\describe{
\item{game}{The non-sequential J! Archive game ID.}
\item{show}{The sequential show number of an episode.}
\item{date}{The air date of an episode.}
}}
\usage{
whatr_info(game)
}
\arguments{
\item{game}{Either a numeric J! Archive game ID or the HTML document object
returned by \code{\link[=whatr_html]{whatr_html()}}, which can take a game ID, air date, show
number, or another HTML document as an input. If an HTML document is not
provided, \code{\link[=whatr_html]{whatr_html()}} will be called.}
}
\value{
A tidy tibble of clue text.
}
\description{
\emph{These} must be given by the contestants in the form of a question in
response to the clues asked.
}
\examples{
whatr_info(game = 6304)
whatr_html(6304) \%>\% whatr_info()
}
