% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whatr-order.R
\name{whatr_order}
\alias{whatr_order}
\title{What is the order?}
\format{A tibble with (usually) 61 rows and 4 variables:
\describe{
\item{round}{The round a clue is chosen.}
\item{col}{The column position left-to-right.}
\item{row}{The row position top-to-bottom.}
}}
\usage{
whatr_order(game)
}
\arguments{
\item{game}{Either a numeric J! Archive game ID or the HTML document object
returned by \code{\link[=whatr_html]{whatr_html()}}, which can take a game ID, air date, show
number, or another HTML document as an input. If an HTML document is not
provided, \code{\link[=whatr_html]{whatr_html()}} will be called.}
}
\value{
A tidy tibble of player info.

A tidy tibble of clue orders.
}
\description{
Scrapes clue order left-to-right, top-to-bottom.
}
