% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whatr-players.R
\name{whatr_players}
\alias{whatr_players}
\title{Who are the players?}
\format{A tibble with 52 rows and 8 variables:
\describe{
\item{first}{The contestant's given name.}
\item{last}{The contestant's surname name.}
\item{occupation}{A short description of what the contestant does.}
\item{city}{The first part of the contestants home, usually a city.}
\item{state}{The second part of the contestants home, usually a state.}
}}
\usage{
whatr_players(html = NULL, game = NULL)
}
\arguments{
\item{html}{An HTML document from \code{\link[=read_game]{read_game()}}.}

\item{game}{The J-Archive! game ID number, possibly from \code{\link[=whatr_id]{whatr_id()}}.}
}
\value{
A tidy tibble of player info.
}
\description{
These individuals compete to score the most points and win the game.
}
\examples{
whatr_players(game = 6304)
read_game(6304) \%>\% whatr_players()
}
