% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whatr-plot.R
\name{whatr_plot}
\alias{whatr_plot}
\title{Visualize a Jeopardy game history}
\usage{
whatr_plot(game = NULL, date = NULL, show = NULL)
}
\arguments{
\item{game_id}{A J-Archive! game ID number}
}
\value{
Return ggplot of a Jeopardy! game score history
}
\description{
Visualize a Jeopardy game history
}
