% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whatr-responses.R
\name{whatr_responses}
\alias{whatr_responses}
\title{Scrape Jeopardy game correct responses}
\usage{
whatr_responses(game = NULL, date = NULL, show = NULL)
}
\arguments{
\item{game}{The J-Archive! game ID number.}

\item{date}{The original date an episode aired.}

\item{show}{The sequential show number.}
}
\value{
A tidy tibble or correct answers.
}
\description{
In this context, the "answers" are the "questions" given by the contentsants
in response to a clue. This function only returns the \emph{correct} response
to a clue, not all responses given by contestants.
}
\examples{
whatr_responses(game = 6304)
}
