% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whatr-summary.R
\name{whatr_summary}
\alias{whatr_summary}
\title{Return a J! Archive game summary}
\usage{
whatr_summary(game = NULL, date = NULL, show = NULL)
}
\arguments{
\item{game}{The J-Archive! game ID number.}

\item{date}{The original date an episode aired.}

\item{show}{The sequential show number.}
}
\value{
The J-Archive! game ID number.
}
\description{
This small tibble summaries the various scoring metrics for each player.
}
\details{
There are three scoring types presented in this summary:
\enumerate{
\item The \href{https://j-archive.com/help.php#coryatscore}{"coryat" score}: a
player's score if all wagering is disregarded. In the Coryat score, there is
no penalty for forced incorrect responses on Daily Doubles, but correct
responses on Daily Doubles earn only the natural values of the clues, and any
gain or loss from the Final Jeopardy! Round is ignored.
\item The final score.
\item The number of clues answered correctly and incorrectly.
}
}
\examples{
whatr_summary(game = 6304)
}
