% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whatr-synopsis.R
\name{whatr_synopsis}
\alias{whatr_synopsis}
\title{What a synopsis?}
\format{A tibble with (usually) 3 row and 5 variables:
\describe{
\item{name}{The contestant's given name.}
\item{coryat}{Score if all wagering is disregarded.}
\item{final}{Final score after Double Jeopardy.}
\item{right}{Number of correct answers.}
\item{wrong}{Number of incorrect answers.}
}}
\usage{
whatr_synopsis(game)
}
\arguments{
\item{game}{One of four types of input, all passed to \code{\link[=whatr_html]{whatr_html()}}:
\enumerate{
\item A numeric game ID.
\item Either a 'showgame' or 'showscores' HTML document.
\item A show number character starting with "#".
\item An air date like "yyyy-mm-dd".
}}
}
\value{
A tidy tibble of clue text.
}
\description{
\emph{This} is an overview containing various summary statistics of a game.
}
\examples{
whatr_synopsis(game = 6304)
}
