% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{synopses}
\alias{synopses}
\title{2019 Jeopardy! game synopses}
\format{
A tibble with 654 rows and 6 variables:
\describe{
\item{name}{The contestant's given name.}
\item{coryat}{Score if all wagering is disregarded.}
\item{final}{Final score after Double Jeopardy.}
\item{right}{Number of correct answers.}
\item{wrong}{Number of incorrect answers.}
}
}
\source{
\url{https://www.j-archive.com/}
}
\usage{
synopses
}
\description{
Synopses of the games in the 2019 season of Jeopardy.
}
\keyword{datasets}
