% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whatr-answers.R
\name{whatr_answers}
\alias{whatr_answers}
\title{What are the answers?}
\format{A tibble with (usually) 61 rows and 5 variables:
\describe{
\item{round}{The round a clue is chosen.}
\item{col}{The column position left-to-right.}
\item{row}{The row position top-to-bottom.}
\item{n}{The order of clue chosen.}
\item{answer}{The \emph{correct} answer to a clue.}
}}
\usage{
whatr_answers(html = NULL, game = NULL)
}
\arguments{
\item{html}{An HTML document from \code{\link[=read_game]{read_game()}}.}

\item{game}{The J-Archive! game ID number, possibly from \code{\link[=whatr_id]{whatr_id()}}.}
}
\value{
A tidy tibble of clue text.
}
\description{
\emph{These} must be given by the contestants in the form of a question in
response to the clues asked.
}
\examples{
whatr_answers(game = 6304)
read_game(6304) \%>\% whatr_answers()
}
