% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whatr-answers.R
\name{whatr_answers}
\alias{whatr_answers}
\title{Scrape Jeopardy game correct responses}
\usage{
whatr_answers(game = NULL, date = NULL, show = NULL)
}
\arguments{
\item{game}{The J-Archive! game ID number.}

\item{date}{The original date an episode aired.}

\item{show}{The sequential show number.}
}
\value{
A tidy tibble or correct answers.
}
\description{
In this context, the "answers" are the "questions" given by the contentsants
in response to a clue. This function only returns the \emph{correct} response
to a clue, not all responses given by contestants.
}
\details{
From the \href{https://j-archive.com/help.php#response}{J! Archive glossary}:
[Responses are] a contestant's "answer" to a clue, phrased in the form of a
question. A correct response results in the addition to a player's score the
value of the clue... an incorrect response results in the subtraction from
the player's score the clue's value (or the wager). A clue may have multiple
different acceptable correct responses or correct response variations, some
of them unanticipated by the writers and judges...
}
\examples{
whatr_answers(game = 6304)
}
