% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whatr-answers.R
\name{whatr_answers}
\alias{whatr_answers}
\title{What are the answers?}
\format{
A tibble with (up to) 61 rows and 5 variables:
\describe{
\item{round}{The round a clue is chosen.}
\item{col}{The column position left-to-right.}
\item{row}{The row position top-to-bottom.}
\item{i}{The order of clue chosen.}
\item{answer}{The \emph{correct} answer to a clue.}
}
}
\usage{
whatr_answers(game)
}
\arguments{
\item{game}{One of four types of input, all passed to \code{\link[=whatr_html]{whatr_html()}}:
\enumerate{
\item A numeric game ID.
\item Either a 'showgame' or 'showscores' HTML document.
\item A show number character starting with "#".
\item An air date like "yyyy-mm-dd".
}}
}
\value{
A tidy tibble of answer text.
}
\description{
\emph{These} must be given by the contestants in the form of a question as a
response to the clues given.
}
\examples{
whatr_answers(game = 6304)
}
