% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whatr-data.R
\name{whatr_board}
\alias{whatr_board}
\title{What is the board?}
\format{A tibble with (usually) 61 rows and 4 variables:
\describe{
\item{round}{The round a clue is chosen.}
\item{col}{The column position left-to-right.}
\item{row}{The row position top-to-bottom.}
\item{n}{The order of clue chosen.}
\item{category}{Category title, often humorous or with instructions.}
\item{clue}{The clue read to the contestants.}
\item{answer}{The \emph{correct} answer to a clue.}
}}
\usage{
whatr_board(html = NULL, game = NULL)
}
\arguments{
\item{html}{An HTML document from \code{\link[=read_game]{read_game()}}.}

\item{game}{The J-Archive! game ID number, possibly from \code{\link[=whatr_id]{whatr_id()}}.}
}
\value{
A tidy tibble of clue text.
}
\description{
\emph{This} grid contains all the categories, clues, and answers in a game.
}
\examples{
whatr_board(game = 6304)
}
