% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whatr-categories.R
\name{whatr_categories}
\alias{whatr_categories}
\title{What are categories?}
\format{A tibble with (usually 13) rows and 3 variables:
\describe{
\item{round}{The round a clue is chosen.}
\item{col}{The column position left-to-right.}
\item{category}{Category title, often humorous or with instructions.}
}}
\usage{
whatr_categories(html = NULL, game = NULL)
}
\arguments{
\item{html}{An HTML document from \code{\link[=read_game]{read_game()}}.}

\item{game}{The J-Archive! game ID number, possibly from \code{\link[=whatr_id]{whatr_id()}}.}
}
\value{
Tidy tibble of category titles.
}
\description{
\emph{These} are collections of five clues related by subject matter.
}
\examples{
whatr_categories(game = 6304)
read_game(6304) \%>\% whatr_categories()
}
