% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whatr-categories.R
\name{whatr_categories}
\alias{whatr_categories}
\title{Scrape Jeopardy game categories}
\usage{
whatr_categories(game = NULL, date = NULL, show = NULL)
}
\arguments{
\item{game}{The J-Archive! game ID number.}

\item{date}{The original date an episode aired.}

\item{show}{The sequential show number.}
}
\value{
A tidy tibble or correct answers.
}
\description{
Categories are collections of five clues, related by subject matter and
arranged in order of difficulty. In the structure of the game board,
categories are the columns.
}
\examples{
whatr_answers(game = 6304)
}
