% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whatr-categories.R
\name{whatr_categories}
\alias{whatr_categories}
\title{Scrape Jeopardy game categories}
\usage{
whatr_categories(game = NULL, date = NULL, show = NULL)
}
\arguments{
\item{game}{The J-Archive! game ID number.}

\item{date}{The original date an episode aired.}

\item{show}{The sequential show number.}
}
\value{
A tibble of categories and positions.
}
\description{
For a given game, return a table of category names and their position on the
board for each round. Categories are assigned to one of six distinct columns
on the game board.
}
\details{
From the \href{https://j-archive.com/help.php#category}{J! Archive glossary}:
[Categories are] a collection of five clues, related by subject matter or
otherwise, (ostensibly) arranged in order of difficulty, and assigned a
(frequently punny) title and, sometimes, special instructions, e.g., for
clues in a category the title of which contains words or letters in quotation
marks, the correct responses will themselves contain (or sometimes begin
with) those words or letters.
}
\examples{
whatr_categories(game = 6304)
}
