% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whatr-categories.R
\name{whatr_categories}
\alias{whatr_categories}
\title{What are categories?}
\format{A tibble with (up to) 13 rows and 3 variables:
\describe{
\item{round}{The round a clue is chosen.}
\item{col}{The column position left-to-right.}
\item{category}{Category title, often humorous or with instructions.}
}}
\usage{
whatr_categories(game)
}
\arguments{
\item{game}{One of four types of input, all passed to \code{\link[=whatr_html]{whatr_html()}}:
\enumerate{
\item A numeric game ID.
\item Either a 'showgame' or 'showscores' HTML document.
\item A show number character starting with "#".
\item An air date like "yyyy-mm-dd".
}}
}
\value{
Tidy tibble of category titles.
}
\description{
\emph{These} are collections of five clues related by subject matter.
}
\examples{
whatr_categories(game = 6304)
}
