% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whatr-clues.R
\name{whatr_clues}
\alias{whatr_clues}
\title{What are the clues?}
\format{A tibble with (usually) 61 rows and 4 variables:
\describe{
\item{round}{The round a clue is chosen.}
\item{col}{The column position left-to-right.}
\item{row}{The row position top-to-bottom.}
\item{clue}{The clue read to the contestants.}
}}
\usage{
whatr_clues(html = NULL, game = NULL)
}
\arguments{
\item{html}{An HTML document from \code{\link[=read_game]{read_game()}}.}

\item{game}{The J-Archive! game ID number, possibly from \code{\link[=whatr_id]{whatr_id()}}.}
}
\value{
A tidy tibble of clue text.
}
\description{
Confusingly, \emph{these} are posed to contestants in the form of an answer.
}
\examples{
whatr_clues(game = 6304)
read_game(6304) \%>\% whatr_clues()
}
