% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whatr-data.R
\name{whatr_data}
\alias{whatr_data}
\title{What are the clues?}
\usage{
whatr_data(game)
}
\arguments{
\item{game}{Either a numeric J! Archive game ID or the HTML document object
returned by \code{\link[=whatr_html]{whatr_html()}}, which can take a game ID, air date, show
number, or another HTML document as an input. If an HTML document is not
provided, \code{\link[=whatr_html]{whatr_html()}} will be called.}
}
\value{
A list of tibbles.
}
\description{
Return
}
\examples{
whatr_data(game = 6304)
}
