% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whatr-board.R, R/whatr-data.R,
%   R/whatr-summary.R
\name{whatr_data}
\alias{whatr_data}
\title{Scrape Jeopardy game meta data}
\usage{
whatr_data(game = NULL, date = NULL, show = NULL)

whatr_data(game = NULL, date = NULL, show = NULL)

whatr_data(game = NULL, date = NULL, show = NULL)
}
\arguments{
\item{game}{A J-Archive! game ID number}

\item{game}{A J-Archive! game ID number}

\item{game}{A J-Archive! game ID number}
}
\value{
Tibble of a Jeopardy! episode categories, questions, and answers

List of a available data on a Jeopardy game

Tibble of Jeopardy! episode summary scores
}
\description{
Scrape Jeopardy game meta data

Scrape Jeopardy game meta data

Scrape Jeopardy game summary scores
}
