% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whatr-data.R
\name{whatr_data}
\alias{whatr_data}
\title{What are the clues?}
\usage{
whatr_data(game)
}
\arguments{
\item{game}{The J-Archive! game ID number, possibly from \code{\link[=whatr_id]{whatr_id()}}.}
}
\value{
A list of tibbles.
}
\description{
Return
}
\examples{
whatr_data(game = 6304)
}
