% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whatr-data.R
\name{whatr_data}
\alias{whatr_data}
\title{What is the board?}
\usage{
whatr_data(game)
}
\arguments{
\item{game}{One of four types of input, all passed to \code{\link[=whatr_html]{whatr_html()}}:
\enumerate{
\item A numeric game ID.
\item Either a 'showgame' or 'showscores' HTML document.
\item A show number character starting with "#".
\item An air date like "yyyy-mm-dd".
}}
}
\value{
A list of tibbles.
}
\description{
\emph{This} is the name for game mechanic displaying all categories and clues.
}
\examples{
whatr_data(game = 6304)
}
