% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whatr-doubles.R
\name{whatr_doubles}
\alias{whatr_doubles}
\title{Get the clue numbers that coorespond to the daily doubles}
\format{a named list
\describe{
\item{single}{The index of the daily double in the first round}
\item{double}{The indices of the two daily doubles in the second round}
}}
\usage{
whatr_doubles(game)
}
\arguments{
\item{game}{Either a numeric J! Archive game ID or the HTML document object
returned by \code{\link[=whatr_html]{whatr_html()}}, which can take a game ID, air date, show
number, or another HTML document as an input. If an HTML document is not
provided, \code{\link[=whatr_html]{whatr_html()}} will be called.}
}
\value{
a list containing the question indices of the daily doubles
in the first and second rounds
}
\description{
Get the clue numbers that coorespond to the daily doubles
}
\examples{
whatr_doubles(game = 6304)
}
