% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whatr-doubles.R
\name{whatr_doubles}
\alias{whatr_doubles}
\title{What are daily doubles?}
\format{a named list
\describe{
\item{round}{The round a clue is chosen.}
\item{col}{The column position left-to-right.}
\item{row}{The row position top-to-bottom.}
\item{n}{The order of clue chosen.}
\item{clue}{The clue read to the contestants.}
\item{score}{The amount won or lost on the wager.}
}}
\usage{
whatr_doubles(game)
}
\arguments{
\item{game}{Either a numeric J! Archive game ID or the HTML document object
returned by \code{\link[=whatr_html]{whatr_html()}}, which can take a game ID, air date, show
number, or another HTML document as an input. If an HTML document is not
provided, \code{\link[=whatr_html]{whatr_html()}} will be called.}
}
\value{
a list containing the question indices of the daily doubles
in the first and second rounds
}
\description{
\emph{These} types of clues have no dollar value. Players wager some of their
score before hearing the clue. In the first round, one such clue is present;
in the second round, there are two.
}
\examples{
whatr_doubles(game = 6304)
whatr_html(6304) \%>\% whatr_doubles()
}
