% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whatr-info.R
\name{whatr_info}
\alias{whatr_info}
\title{What is the information?}
\format{A tibble with 1 row and 3 variables:
\describe{
\item{game}{The non-sequential J! Archive game ID.}
\item{show}{The sequential show number of an episode.}
\item{date}{The air date of an episode.}
}}
\usage{
whatr_info(game)
}
\arguments{
\item{game}{One of four types of input, all passed to \code{\link[=whatr_html]{whatr_html()}}:
\enumerate{
\item A numeric game ID.
\item Either a 'showgame' or 'showscores' HTML document.
\item A show number character with "#".
\item An air date like "yyyy-mm-dd".
}}
}
\value{
A tidy tibble of clue text.
}
\description{
\emph{These} must be given by the contestants in the form of a question in
response to the clues asked.
}
\examples{
whatr_info(game = 6304)
whatr_html(6304) \%>\% whatr_info()
}
