% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whatr-info.R
\name{whatr_info}
\alias{whatr_info}
\title{What is the information?}
\format{A tibble with 1 row and 3 variables:
\describe{
\item{game}{The non-sequential J! Archive game ID.}
\item{show}{The sequential show number of an episode.}
\item{date}{The air date of an episode.}
}}
\usage{
whatr_info(html = NULL, game = NULL)
}
\arguments{
\item{html}{An HTML document from \code{\link[=read_game]{read_game()}}.}

\item{game}{The J-Archive! game ID number, possibly from \code{\link[=whatr_id]{whatr_id()}}.}
}
\value{
A tidy tibble of clue text.
}
\description{
\emph{These} must be given by the contestants in the form of a question in
response to the clues asked.
}
\examples{
whatr_info(game = 6304)
read_game(6304) \%>\% whatr_info()
}
