% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whatr-info.R
\name{whatr_info}
\alias{whatr_info}
\title{Scrape Jeopardy game air data}
\usage{
whatr_info(game = NULL, date = NULL, show = NULL)
}
\arguments{
\item{game}{A J-Archive! game ID number}
}
\value{
Tibble of a Jeopardy! episode air information
}
\description{
Scrape Jeopardy game air data
}
