% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whatr-order.R
\name{whatr_order}
\alias{whatr_order}
\title{Scrape Jeopardy game clues order}
\usage{
whatr_order(game = NULL, date = NULL, show = NULL)
}
\arguments{
\item{game}{The J-Archive! game ID number.}

\item{date}{The original date an episode aired.}

\item{show}{The sequential show number.}
}
\value{
A tidy tibble of clue orders.
}
\description{
Contestants are free to chose clues from the board in whatever order they
want. This function returns the position of each clue on the board and the
order in which they are chosen.
}
