% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whatr-players.R
\name{whatr_players}
\alias{whatr_players}
\title{Who are the players?}
\format{A tibble with 52 rows and 8 variables:
\describe{
\item{first}{The contestant's given name.}
\item{last}{The contestant's surname name.}
\item{occupation}{A short description of what the contestant does.}
\item{city}{The first part of the contestants home, usually a city.}
\item{state}{The second part of the contestants home, usually a state.}
}}
\usage{
whatr_players(game)
}
\arguments{
\item{game}{One of four types of input, all passed to \code{\link[=whatr_html]{whatr_html()}}:
\enumerate{
\item A numeric game ID.
\item Either a 'showgame' or 'showscores' HTML document.
\item A show number character with "#".
\item An air date like "yyyy-mm-dd".
}}
}
\value{
A tidy tibble of player info.
}
\description{
These individuals compete to score the most points and win the game.
}
\examples{
whatr_players(game = 6304)
whatr_html(6304) \%>\% whatr_players()
}
