% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whatr-players.R
\name{whatr_players}
\alias{whatr_players}
\title{Who are the players?}
\format{A tibble with 52 rows and 8 variables:
\describe{
\item{first}{The contestant's given name.}
\item{last}{The contestant's surname name.}
\item{occupation}{A short description of what the contestant does.}
\item{city}{The first part of the contestants home, usually a city.}
\item{state}{The second part of the contestants home, usually a state.}
}}
\usage{
whatr_players(game)
}
\arguments{
\item{game}{Either a numeric J! Archive game ID or the HTML document object
returned by \code{\link[=whatr_html]{whatr_html()}}, which can take a game ID, air date, show
number, or another HTML document as an input. If an HTML document is not
provided, \code{\link[=whatr_html]{whatr_html()}} will be called.}
}
\value{
A tidy tibble of player info.
}
\description{
These individuals compete to score the most points and win the game.
}
\examples{
whatr_players(game = 6304)
whatr_html(6304) \%>\% whatr_players()
}
