% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whatr-players.R
\name{whatr_players}
\alias{whatr_players}
\title{Scrape Jeopardy game meta data}
\usage{
whatr_players(game = NULL, date = NULL, show = NULL)
}
\arguments{
\item{game_id}{A J-Archive! game ID number}
}
\value{
Tibble of a Jeopardy game player names
}
\description{
Scrape Jeopardy game meta data
}
