% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whatr-players.R
\name{whatr_players}
\alias{whatr_players}
\title{Who are the players?}
\usage{
whatr_players(html = NULL, game = NULL)
}
\arguments{
\item{html}{An HTML document from \code{\link[=read_scores]{read_scores()}}.}

\item{game}{The J-Archive! game ID number.}
}
\value{
A tidy tibble of player info.
}
\description{
These individuals compete to score the most points and win the game.
}
\examples{
whatr_players(game = 6304)
read_game(6304) \%>\% whatr_players()
}
