% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whatr-plot.R
\name{whatr_plot}
\alias{whatr_plot}
\title{What is a plot?}
\usage{
whatr_plot(game)
}
\arguments{
\item{game}{One of four types of input, all passed to \code{\link[=whatr_html]{whatr_html()}}:
\enumerate{
\item A numeric game ID.
\item Either a 'showgame' or 'showscores' HTML document.
\item A show number character starting with "#".
\item An air date like "yyyy-mm-dd".
}}
}
\value{
A ggplot object.
}
\description{
\emph{This} type of graphic presents data in a visual manner.
}
\examples{
whatr_plot(game = 6304)
}
