% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whatr-scores.R
\name{whatr_scores}
\alias{whatr_scores}
\title{Scrape Jeopardy game score history}
\usage{
whatr_scores(game = NULL, date = NULL, show = NULL)
}
\arguments{
\item{game}{The J-Archive! game ID number.}

\item{date}{The original date an episode aired.}

\item{show}{The sequential show number.}
}
\value{
Tidy tibble of an episode score history.
}
\description{
Return the tidy tibble of contestant scores over the course of the game.
The tibble indicates what questions were answered correctly and the change
in score from each answer.
}
\examples{
whatr_scores(game = 6304)
whatr_scores(date = "2019-06-03")
whatr_scores(show = 8006)
}
\concept{J! scrapers}
