% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whatr-scores.R
\name{whatr_scores}
\alias{whatr_scores}
\title{What are player scores?}
\format{A tibble with 52 rows and 8 variables:
\describe{
\item{round}{The round a clue is chosen.}
\item{n}{The order of clue chosen.}
\item{name}{First name of player responding.}
\item{score}{Change in score from this clue.}
\item{double}{Is the clue a daily double.}
}}
\usage{
whatr_scores(game)
}
\arguments{
\item{game}{Either a numeric J! Archive game ID or the HTML document object
returned by \code{\link[=whatr_html]{whatr_html()}}, which can take a game ID, air date, show
number, or another HTML document as an input. If an HTML document is not
provided, \code{\link[=whatr_html]{whatr_html()}} will be called.}
}
\value{
Tidy tibble of clue scores.
}
\description{
This data frame has data on the performance of each contestant.
}
\examples{
whatr_scores(game = 6304)
whatr_html(6304, "showscores") \%>\% whatr_scores()
}
