% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whatr-scores.R
\name{whatr_scores}
\alias{whatr_scores}
\title{What are player scores?}
\format{A tibble with 52 rows and 8 variables:
\describe{
\item{round}{The round a clue is chosen.}
\item{n}{The order of clue chosen.}
\item{name}{First name of player responding.}
\item{score}{Change in score from this clue.}
\item{double}{Is the clue a daily double.}
}}
\usage{
whatr_scores(html = NULL, game = NULL)
}
\arguments{
\item{html}{An HTML document from \code{\link[=read_scores]{read_scores()}}.}

\item{game}{The J-Archive! game ID number, possibly from \code{\link[=whatr_id]{whatr_id()}}.}
}
\value{
Tidy tibble of clue scores.
}
\description{
This data frame has data on the performance of each contestant.
}
\examples{
whatr_scores(game = 6304)
read_scores(6304) \%>\% whatr_scores()
}
