% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whatr-summary.R
\name{whatr_summary}
\alias{whatr_summary}
\title{What is the information?}
\format{A tibble with (usually) 3 row and 5 variables:
\describe{
\item{name}{The contestant's given name.}
\item{coryat}{Score if all wagering is disregarded.}
\item{final}{Final score after Double Jeopardy.}
\item{right}{Number of correct answers.}
\item{wrong}{Number of incorrect answers.}
}}
\usage{
whatr_summary(html = NULL, game = NULL)
}
\arguments{
\item{html}{An HTML document from \code{\link[=read_game]{read_game()}}.}

\item{game}{The J-Archive! game ID number, possibly from \code{\link[=whatr_id]{whatr_id()}}.}
}
\value{
A tidy tibble of clue text.
}
\description{
\emph{These} must be given by the contestants in the form of a question in
response to the clues asked.
}
\examples{
whatr_summary(game = 6304)
}
